requirements.txt
Flask==3.0.0
python-dotenv==2.1.0
requests==2.31.0
mariadb==1.1.10
